+++ src/main/java/insane96mcp/survivalreimagined/mixin/EnchantmentMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.survivalreimagined.mixin;
 
-import insane96mcp.survivalreimagined.module.experience.feature.Enchantments;
+import insane96mcp.survivalreimagined.module.experience.feature.EnchantmentsFeature;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.ArrowInfiniteEnchantment;
 import net.minecraft.world.item.enchantment.Enchantment;
@@ -13,13 +13,13 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class EnchantmentMixin {
 	@Inject(at = @At("RETURN"), method = "canEnchant", cancellable = true)
 	private void canApply(ItemStack stack, CallbackInfoReturnable<Boolean> callback) {
-		if (Enchantments.disableEnchantment((Enchantment) (Object) this))
+		if (EnchantmentsFeature.disableEnchantment((Enchantment) (Object) this))
 			callback.setReturnValue(false);
 	}
 
 	@Inject(at = @At("RETURN"), method = "canApplyAtEnchantingTable", cancellable = true, remap = false)
 	private void canApplyAtEnchantingTable(ItemStack stack, CallbackInfoReturnable<Boolean> callback) {
-		if (Enchantments.disableEnchantment((Enchantment) (Object) this))
+		if (EnchantmentsFeature.disableEnchantment((Enchantment) (Object) this))
 			callback.setReturnValue(false);
 	}